services/trade_producer/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Taras <tarasenya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
quixstreams = "^2.5.1"
websocket-client = "^1.8.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]